+++ src/MailcoachApiTransport.php
@@ -9,6 +9,7 @@ use Spatie\MailcoachMailer\Exceptions\NoHostSet;
 use Spatie\MailcoachMailer\Exceptions\NotAllowedToSendMail;
 use Spatie\MailcoachMailer\Headers\FakeHeader;
 use Spatie\MailcoachMailer\Headers\GoogleAnalyticsCampaignHeader;
+use Spatie\MailcoachMailer\Headers\GoogleAnalyticsDomainsHeader;
 use Spatie\MailcoachMailer\Headers\MailerHeader;
 use Spatie\MailcoachMailer\Headers\ReplacementHeader;
 use Spatie\MailcoachMailer\Headers\StoreContentHeader;
@@ -126,6 +127,10 @@ class MailcoachApiTransport extends AbstractApiTransport
 
             if ($header instanceof GoogleAnalyticsCampaignHeader) {
                 $payload['google_analytics_campaign'] = $header->getValue();
+            }
+
+            if ($header instanceof GoogleAnalyticsDomainsHeader) {
+                $payload['google_analytics_domains'] = json_decode($header->getValue(), true);
             }
         }